Update patch set 1
Patch Set 1:

I was experimenting with "Remove device" button position and in my opinion, the header is where it fits best.

I do, however, agree that it could be not obvious for the first time. Maybe add some tooltip for it or even replace with a simple text link in brackets? Like this: "Sony C6503 (remove device)".

What do you think?

Patch-set: 1