Update patch set 2
Patch Set 2:

@Chris, might I suggest an enhancement for the site?

When clicking "Remote Wipe", you are asked to provide your password. After entering it, the site immediately initiates the wipe!!

Would it not be better to add an extra Are you sure? warning?

It scared the hell out of me.


I was just testing the functionality of the button and the translation. Luckily I was able to put my phone in Flight Mode

Patch-set: 2