Update patch set 2
Patch Set 2:

This is not the place to report or discuss issues, though I have made a note of them.

@Marco - You can edit the language in app/js/cmaccount/l10n.js.

As for the wipe confirmation, there is no need to add a confirmation screen to the flow.  We can just use the password prompt as the "Are you sure?" verification.  The act of entering your password should be confirmation, we just need to display some string on that page that warns you.

Patch-set: 2