Update patch set 2
Patch Set 2:

Not working with Chrome on OS X, though it could be my environment.  I'll play around with it a little bit.

We can style the tooltips to match our color scheme a bit better.  The nice thing about bootstrap tooltips is that the feedback is immediate, instead of hovering over the X until the browser displays the title tooltip.  That's my only gripe with not having them.

Patch-set: 2