Update patch set 2
Patch Set 2:

Honestly, I don't think "immediate" tooltip is a good thing.

Let me try explain this a little bit (warning, bad English!): the purpose of a tooltip is to give user a hint on item he is pointing at (or touching), but only if the item action is not obvious for him. That's how it works in every operating system (including Android: to get a tooltip for ActionBar item, you should hold a finger on it for a while).

"X" sign should be pretty obvious. Even if it's not, you can hover the cursor on it, read "Remove device" and _remember_ it. You won't be distracted by tooltips constantly popping out when you're just moving the cursor on the page ("hey, I already know "X" is to remove device, why show it again?").

I'll reboot into Windows and check if tooltips work there, just a sec.

Patch-set: 2