Update patch set 3
Patch Set 3:

I wouldn't describe the tooltip as buggy.  I'm not experiencing the same issue on Chrome on OS X.

I still don't care for this patch, no offense.  I don't think the button is as obvious as it was when it was near all of the other buttons.  The only way I am personally willing to accept this is with the bootstrap tooltips.

Patch-set: 3