Update patch set 3
Patch Set 3:

I honestly don't agree with this patch.  We are talking about a destructive operation.  A 'X' button in the corner is ambiguous as to it's intention at best. 

"Does it dismiss the device?"
"Does it remove the device completely?"
"If I 'Close' this device, how do i get it back?"

The intention of the original button was clear.

If this change is accepted, we absolutely must use a bootstrap tooltip.  Not only does it keep a consistent look and feel with the rest of the site, the instant nature of this tooltip provides valuable information to this otherwise ambiguous operation.

Additionally, using https://www.google.com/android/devicemanager as an example isn't exactly fitting for this because of how they are used.  The buttons with destructive side effects have text under them to describe their operations to begin with.  The other buttons have clear outcomes because of the icons used.

Patch-set: 3