Update patch set 1
Patch Set 1:

We have been getting a ton of emails from users that do not seem to understand that after verifying their email address, the link expires.  Email verification gets us nothing.  It does not prevent spam, as spammers do not have an incentive to create an account (ie: posting on a comment somewhere).

Best course of action is to just remove it for now.  Server side has already been updated.

Patch-set: 1