Update patch set 3
Patch Set 3:

I'm sure we could, we should be able to override any bootstrap styling that is fixing a specific max-width.

However, I'm still not entirely okay with moving the remove button.  It is very clear where it is right now.  If we get a bunch of +1s, i'll do it, but most of the people I have spoken to agree that it is more clear where it is now.

Patch-set: 3